Use Entropy || PrivateKey as Key secret
The 'encrypted' flag is now not being passed to external anymore, and is replaced by automatic detection of the stored secret length. The type of a Key is now determined by the type of the secret, but is still stored in the DB and cookie, because the type can only be read when the secret is decrypted.