Dedupe BTC key paths to prevent "No inputs were signed" error
When multiple inputs of one address are used for a swap, its keyPair is derived multiple times as well, and the private key is stored for the swap iframe multiple times. There it is then tried to sign the inputs multiple times, failing after the first round, as the inputs to sign are already signed.