BrowserStack local iOS testing compatibility
On iOS, BrowserStack routes localhost connections for testing of local code
through an actual domain (bs-local.com) because iOS blocks localhost connections
to content that's not sitting on the phone.
This commit adds changes to treat bs-local.com just like localhost, such that
a locally served dev server can be tested with BrowserStack.