Enable Cookie Monster service worker
Strip cookies from requests via a service worker to avoid that the cookies
are sent to the server, which are only meant to store data locally. The
Cookie Monster service worker had already been several years ago, however
back then it couldn't be enabled, because Safari had issues with multiple
service workers in the same tab in the top-level window and iframes. These
issues do not seem to exist anymore and the service worker is now safe to
be enabled.