Overlay: remove now unused Overlay component
The Overlay component is not used in the create flow or anywhere else anymore since
USDC support was merged (PR #465), and is incompatible with the changes in the iOS
pointer-events bug fix (PR #470), where pointer-events were enabled on the overlay-frame
when they shouldn't. The incompatibility does not show directly anymore since PR #465 was
merged, but showed on the master branch before PR #465 that PR #470 was originally targeting.