Allow SignTransaction to sign HTLC redeem transactions
Which is used by the manual refund flow for failed NIM swaps.

Before this change, only _basic_ transactions were supported by SignTransaction, but a HTLC refund tx is an extended transaction from a HTLC account type, which was not covered by the cases.

This changes it to use the already-parsed `transaction` from the request object, which allows signing of all transactions that the request accepts in the first place.

We also don't worry about any proof prefix bytes here, as those are added in the Hub when the result of this Keyguard request is received. The signature will still be correct, because a signature is not made over the proof prefix bytes.