Fix calculated validityStartHeight of swap NIM refund transactions
ValidityStartHeight is still a block number in PoS, not a timestamp.

This is not exact, unfortunately. But hopefully close enough. The calculated VSH is always smaller than it should be (blocks are faster than 1/second on average, time taken between request creation and this code is subtracted as well). But the watchtower only send the refund transaction once the timeout _timestamp_ of the swap has passed, at which point the transaction will still be valid.