Key: security fix for RSA key seed generation
Previoulsy, the bytes have been interpreted as UTF8 bytes and encoded to a Javascript string,
which is UTF16 encoded. Up to 4 UTF8 bytes can represent a single unicode symbol, which is then
encoded as a single UTF16 character or a UTF16 surrogate pair. The following re-encodings are
possible:
1 UTF8 byte -> 1 UTF16 character
2 UTF8 bytes -> 1 UTF16 character
3 UTF8 bytes -> 1 UTF16 character
4 UTF8 bytes -> 2 UTF16 characters (surrogate pair)

This means that the resulting string could be as short as a third of the length of the input byte
array. As node-forge expects a binary string as input, where each character is interpreted as a
single byte, overall in the worst case, two thirds of the original data were discarded and an
exponentially weaker entropy was used.

This fix is a breaking change, as the resulting RSA key will be different. Therefore, users who
already have a stored RSA key need to logout to clear their old RSA key (or we decide to rename the
property in the store). Any app that already got the old RSA public key via a Connect request must
discard its known RSA public keys and associated encrypted data, and connect the account again to
get the new RSA public key.