Key,RSAKeysIframe: security fix for postMessage communication with RSA iframe
Reject any messages in communication between the RSA iframe and main window, which are not
from / to the intended sender / recipient. Otherwise, the following attack was possible:
A malicious site starts a Connect request in an iframe (via Hub, which then redirects to the
Keyguard). Then, when the malicious site suspects that the Keyguard is waiting for the RSA key
calculation, it itself sends a postmessage to the popup with an RSA key known to the attacker,
which the Keyguard would willingly store and continue to use, also for other websites.