Connect: provide more accurate information on exposed addresses
Previously, the UI wrongly claimed, that always the first address is exposed, which
is only the most common case. In fact though, an app can request arbitrary addresses
and an arbitrary number of addresses. In such case, the UI now shows the number of
requested addresses and their derivation paths.