Connect: security fix for blind signing of arbitrary messages
The Connect request blindly signed any arbitrary challenge as a valid Nimiq message
signature. I.e. an attacker could trick the user into unknowingly signing messages
with the only restriction enforced by request parsing being that the message must be
textual, which is the most common case for messages to be approved by the user. This
could be abused for impersonating the user and for example authenticating at staking
pool websites as the user.

To avoid this, a distinct prefix separate to the Nimiq message signature prefix is
now used.

It's a breaking change, which also requires adaptions to the signature verification
in apps using the Connect request.

This commit also improves the way message prefixes are shared between the Keyguard
client and the main Keyguard source code by redefining the message prefixes in a
type-checked fashion in the main code, instead of the client having to import it as
a javascript file and exposing the value in a hacky global variable on window.